Make AbstractProjectCodeModel just be CodeModel
Now that there's nothing being brought by the inheritence model, we can
seal the type.
