+++ src/VisualStudio/Core/Impl/ProjectSystem/CPS/CPSProject_IProjectCodeModelProvider.cs
@@ -1,21 +1,23 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System;
 using System.Threading;
 using Microsoft.VisualStudio.LanguageServices.Implementation.CodeModel;
+using Microsoft.VisualStudio.LanguageServices.Implementation.ProjectSystem.Extensions;
 
 namespace Microsoft.VisualStudio.LanguageServices.Implementation.ProjectSystem.CPS
 {
     internal sealed partial class CPSProject : IProjectCodeModelProvider
     {
-        private AbstractProjectCodeModel _projectCodeModel;
+        private ProjectCodeModel _projectCodeModel;
 
-        public AbstractProjectCodeModel ProjectCodeModel
+        public ProjectCodeModel ProjectCodeModel
         {
             get
             {
                 if (_projectCodeModel == null && this.Workspace != null)
                 {
-                    Interlocked.CompareExchange(ref _projectCodeModel, new CPSProjectCodeModel(this, (VisualStudioWorkspaceImpl)this.Workspace, ServiceProvider), null);
+                    Interlocked.CompareExchange(ref _projectCodeModel, new ProjectCodeModel(this.Id, new CPSCodeModelInstanceFactory(this), (VisualStudioWorkspaceImpl)this.Workspace, ServiceProvider), null);
                 }
 
                 return _projectCodeModel;
@@ -32,12 +34,49 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.ProjectSystem.C
 
         public EnvDTE.CodeModel GetCodeModel(EnvDTE.Project parent)
         {
-            return ((CPSProjectCodeModel)ProjectCodeModel)?.GetCodeModel(parent);
+            return ProjectCodeModel.GetOrCreateRootCodeModel(parent);
         }
 
         public EnvDTE.FileCodeModel GetFileCodeModel(EnvDTE.ProjectItem item)
         {
-            return ((CPSProjectCodeModel)ProjectCodeModel)?.GetFileCodeModel(item);
+            if (!item.TryGetFullPath(out var filePath))
+            {
+                return null;
+            }
+
+            return ProjectCodeModel.GetOrCreateFileCodeModel(filePath, item).Handle;
+        }
+
+        private class CPSCodeModelInstanceFactory : ICodeModelInstanceFactory
+        {
+            private CPSProject _project;
+
+            public CPSCodeModelInstanceFactory(CPSProject project)
+            {
+                _project = project;
+            }
+
+            EnvDTE.FileCodeModel ICodeModelInstanceFactory.TryCreateFileCodeModelThroughProjectSystem(string filePath)
+            {
+                var projectItem = GetProjectItem(filePath);
+                if (projectItem == null)
+                {
+                    return null;
+                }
+
+                return _project.ProjectCodeModel.GetOrCreateFileCodeModel(filePath, projectItem).Handle;
+            }
+
+            private EnvDTE.ProjectItem GetProjectItem(string filePath)
+            {
+                var dteProject = ((VisualStudioWorkspaceImpl)_project.Workspace).TryGetDTEProject(_project.Id);
+                if (dteProject == null)
+                {
+                    return null;
+                }
+
+                return dteProject.FindItemByPath(filePath, StringComparer.OrdinalIgnoreCase);
+            }
         }
     }
 }
